Add files via upload
This is a draft of what ive worked so far for my part (Inventory Record Management). Just uploading for reference just in case. Main issue is the main code logic, gui and database arent connected yet. Feel free to redo if this is wrong or let me know so i can revise as needed skskskksks (I had to download mysql-connector-j-9.1.0.jar and apache tomcat for this (but not yet working AHHHH)